Improves weekly snippets.
This has been iffy for some time but I found a separate advanced search functionality in GitHub which works across repositories. With this I was able to make a few improvements:
* The snippet does not require a repository name, it searches all repositories.
* The snippet now searches for PRs, rather than commits.
* Fixed the date which needed to be zero-padded for GitHub to use it correctly.

I'm still a little unclear about the `created:` atom, as I thing I want some kind of `closedAt:` atom which doesn't seem to exist. This seems to work as I want it to, but there may be some edge cases where this doesn't work out and may need to be revisited.