Adds "Jump to GitHub Issue" bookmarklet.
This is a bookmarklet which prompts the user for a repository and issue/PR number then navigates to that page. It's particularly useful when in a meeting and someone says they're looking at issue XXX, without requiring them to send a link.

Trickiest part was making this work when normally clicked to navigate the current tab vs opening a new tab. Had to hook into the visibility API to make this work. May be worth backporting this solution to the Weekly Summary bookmarklet.

Most interesting thing is that the file overview comment mentions how the script needs to be contained in:

```html
<html>
  <head>
    <title>Jump to GitHub Issue</title>
  </head>
  <body>
    <script>${snippet}</script>
  </body>
</html>
```

This is necessary to avoid injecting JavaScript onto the `chrome://newtab` page, which is explicitly blocked. The title is a simple convenience, however we can't actually include the characters `</script>` in this bookmarklet (even in comments), because the page will interpret that as the end of the JavaScript. So the comment escapes this tag as `<\/script>` but tries to call this out so users don't just copy paste it.